# bo/specs/001-checkbox-based-permission/contracts/permission-api.openapi.yaml
openapi: 3.0.3
info:
  title: Checkbox Permission System API (No Admin Plugin)
  description: API contracts for permission management without Better Auth admin plugin
  version: 1.0.0
  contact:
    name: PlayHard BO Team
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /organizations/{organizationId}/users:
    get:
      summary: List organization users (replacement for /api/admin/users)
      tags: [Admin Migration]
      parameters:
        - name: organizationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Users list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/OrganizationUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  schemas:
...
    OrganizationUser:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: string
        memberRole:
          type: string
        lastActive:
          type: string
          format: date-time
      required: [id, email]
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
      required: [error]
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    BetterAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Better Auth session cookie
security:
  - BetterAuth: []
...